// server/src/catalog/models/catalog-filter.dto.ts
// catalog-filter.dto.ts

export class CatalogFilterDto {
    name?: string;
    price?: number;
    styles?: string;
    rooms?: string;
    width?: number;
    height?: number;
    depth?: number;
    colors?: string;
    object_id?: string;
    model_id?: number;
    active?: boolean;
    company?: number;
    company_name?: string;
  }
  